Move doppler.go into main.go
